backend_2/requirements.txt
fastapi
uvicorn
paramiko
python-dotenv
pydantic-settings